Ensure cleanServerGlobals only run once
This is preparation for dependencies loading this library via Composer.
Previously they will require_once ('pureContent.php') which means pureContent::cleanServerGlobals (); is always run.
With composer they will have to replace that require_once with a cleanServerGlobals () call directly, to replicate it being run automatically.
So we need to avoid it being run twice.
Later, once all callers run it directly, the loading in this class definition should be removed.